+++ example/lib/advanced_example.dart
@@ -35,14 +35,16 @@ class _AdvancedExamplePageState extends State<AdvancedExamplePage> {
     dio.interceptors.clear();
     dio.interceptors.add(
       FixturesInterceptor(
-        sources: [
-          HttpFileFixtureSource(),
-          OpenApiFixtureSource(specPath: 'assets/fixtures/openapi.json'),
-        ],
-        dataSelectorView: FixturesDialogView(
-          contextProvider: () => widget.navigatorKey.currentContext!,
+        pipeline: FixturePipeline(
+          source: HttpFixtureSources([
+            HttpFileFixtureSource(),
+            OpenApiFixtureSource(specPath: 'assets/fixtures/openapi.json'),
+          ]),
+          selector: _getDataSelectorType(),
+          view: FixturesDialogView(
+            contextProvider: () => widget.navigatorKey.currentContext!,
+          ),
         ),
-        dataSelector: _getDataSelectorType(),
       ),
     );
   }